requirements.txt
langchain
langchain-groq
python-dotenv
langchain-community
ddgs
flask
flask-cors